third_party/blink/web_tests/external/wpt/svg/interact/scripted/tabindex-focus-flag-expected.txt

This is a testharness.js-based test.
[FAIL] The document root element should be focusable.
  assert_equals: expected Element node <svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://w... but got null
[FAIL] rect[focusable=true] should be focusable.
  assert_equals: expected Element node <rect focusable="true"></rect> but got null
[FAIL] a[focusable=false] should not be focusable.
  assert_not_equals: got disallowed value Element node <a href="" focusable="false"></a>
[FAIL] iframe should be focusable.
  assert_equals: expected Element node <iframe src="resources/blank.htm"></iframe> but got Element node <a href="" focusable="false"></a>
[FAIL] canvas[tabindex] should be focusable.
  assert_equals: expected Element node <canvas tabindex="0"></canvas> but got Element node <video controls="controls"></video>
[FAIL] mesh[tabindex] should be focusable.
  assert_equals: expected Element node <mesh tabindex="0"></mesh> but got Element node <line tabindex="0"></line>
[FAIL] unknown[tabindex] should be focusable.
  assert_equals: expected Element node <unknown tabindex="0"></unknown> but got Element node <tspan tabindex="0"></tspan>
Harness: the test ran to completion.
